requirements.txt
flask
flask-cors
pymongo
requests
python-dotenv
gspread
